Hopefully fix compile issue, and hopefully get rid of nearly irrelevant compile warnings.
Added Wt as "SYSTEM" includes in CMake, to hopefully get rid of warnings caused by those headers, that we cant really do much about at the moment.
Also hopefully fix compile issue in SimpleActivityCalc.

+++ src/RelActAutoGui.cpp
@@ -3319,7 +3319,7 @@ void RelActAutoGui::handleRemovePartOfEnergyRange( Wt::WWidget *w,
     return;
   }
   
-  const int orig_w_index = pos - begin( kids );
+  const int orig_w_index = static_cast<int>( pos - begin(kids) );
   RelActAutoGuiEnergyRange *range = dynamic_cast<RelActAutoGuiEnergyRange *>( w );
   assert( range );
   if( !range )
